Upped the player speed a little and hid the level geometry.

+++ rust/src/tiled_geometry.rs
@@ -114,7 +114,7 @@ impl TiledGeometry {
 		self.collision_rects = simplify_rects(&mut self.collision_rects);
 		self.tracks = simplify_rects(&mut self.tracks);
 		// For debugging: draw all the rectangles.
-		{
+		if false {
 			let mut editor = self.debug_buffer.make_editor();
 			editor.clear();
 			{
